Cleanup tests
The tests here were getting messy, this separates things into
logical groupings of functionality, including a file specifically
for testing kommander, and one for testing thanos.

Also marked some functionality that belongs upstream.
