+++ test/unhandled_addons_test.go
@@ -1,0 +1,55 @@
+package test
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/mesosphere/kubeaddons/pkg/api/v1beta1"
+	"github.com/mesosphere/kubeaddons/pkg/repositories/local"
+)
+
+// TODO - this needs to be moved upstream to the test framework
+func TestValidateUnhandledAddons(t *testing.T) {
+	unhandled, err := findUnhandled()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if len(unhandled) != 0 {
+		names := make([]string, len(unhandled))
+		for _, addon := range unhandled {
+			names = append(names, addon.GetName())
+		}
+		t.Fatal(fmt.Errorf("the following addons are not handled as part of a testing group: %+v", names))
+	}
+}
+
+// TODO - this needs to be moved upstream to the test framework
+func findUnhandled() ([]v1beta1.AddonInterface, error) {
+	var unhandled []v1beta1.AddonInterface
+	repo, err := local.NewRepository("base", "../addons")
+	if err != nil {
+		return unhandled, err
+	}
+	addons, err := repo.ListAddons()
+	if err != nil {
+		return unhandled, err
+	}
+
+	for _, revisions := range addons {
+		addon := revisions[0]
+		found := false
+		for _, v := range groups {
+			for _, r := range v {
+				if r.GetName() == addon.GetName() {
+					found = true
+				}
+			}
+		}
+		if !found {
+			unhandled = append(unhandled, addon)
+		}
+	}
+
+	return unhandled, nil
+}
